requirements.txt
geopandas
shapely
rasterio
python-dotenv
xarray
netCDF4
pillow
matplotlib
requests
copernicusmarine
h5py